% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_import.R
\name{importCounts}
\alias{importCounts}
\title{Import count files}
\usage{
importCounts(count_files, skip = 0)
}
\arguments{
\item{count_files}{Character vector with absolute paths to count files}

\item{skip}{Integer for defining how many lines to skip in the count files}
}
\value{
A data.frame consisting of ID by rownames and count values by columns.
Each column are named after the filename, without their file extensions.
}
\description{
Automatically imports count files and converts 0 values to NA,
for easier zero-inflation removal.
}
\details{
Note all input files, must contain the same Gene or Transcript ID's. These must likewise be ordered identically.
}
