% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_trim.R
\name{trimCounts}
\alias{trimCounts}
\title{Trim count table
Remove columns which contains 0-counts amounting up to the set threshold in percentages.}
\usage{
trimCounts(data, threshold, silent = FALSE)
}
\arguments{
\item{data}{data.frame as provided by {importCounts}}

\item{threshold}{numeric Upper threshold in percent for how many zero counts are allowed in a sample (between 1 and 100).}

\item{silent}{logical to ignore system warnings.}
}
\value{

}
\description{
Trim count table
Remove columns which contains 0-counts amounting up to the set threshold in percentages.
}
